refactor: Mapping 을 Controller 로 이름 변경
