+++ tomcat/src/main/java/org/apache/coyote/handler/controller/HomePageController.java
@@ -1,6 +1,6 @@
-package org.apache.coyote.handler.mapping.register;
+package org.apache.coyote.handler.controller;
 
-import org.apache.coyote.handler.AbstractController;
+import org.apache.coyote.handler.RequestController;
 import org.apache.coyote.http.common.ContentType;
 import org.apache.coyote.http.common.HttpBody;
 import org.apache.coyote.http.request.HttpRequest;
@@ -8,16 +8,16 @@ import org.apache.coyote.http.response.HttpResponse;
 import org.apache.coyote.http.response.StatusCode;
 import org.apache.coyote.http.response.StatusLine;
 
-import static org.apache.coyote.handler.mapping.Path.REGISTER;
 import static org.apache.coyote.http.common.HttpHeader.CONTENT_TYPE;
 
-public class RegisterPageMapping extends AbstractController {
+public class HomePageController extends RequestController {
 
-    private static final String TARGET_URI = "register";
+    private static final String HOME_MESSAGE = "Hello world!";
+    private static final String TARGET_URI = "/";
 
     @Override
     public boolean supports(final HttpRequest httpRequest) {
-        return httpRequest.isGetRequest() && httpRequest.containsRequestUri(TARGET_URI);
+        return TARGET_URI.equals(httpRequest.getRequestUri().getRequestUri());
     }
 
     @Override
@@ -29,6 +29,6 @@ public class RegisterPageMapping extends AbstractController {
     protected void doGet(final HttpRequest request, final HttpResponse response) throws Exception {
         response.changeStatusLine(StatusLine.from(StatusCode.OK));
         response.addHeader(CONTENT_TYPE, ContentType.HTML.getValue());
-        response.changeBody(HttpBody.file(REGISTER.getPath()));
+        response.changeBody(new HttpBody(HOME_MESSAGE));
     }
 }
