+++ tomcat/src/main/java/org/apache/coyote/handler/controller/login/LoginController.java
@@ -1,8 +1,8 @@
-package org.apache.coyote.handler.mapping.login;
+package org.apache.coyote.handler.controller.login;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.handler.AbstractController;
+import org.apache.coyote.handler.RequestController;
 import org.apache.coyote.http.LoginManager;
 import org.apache.coyote.http.common.HttpBody;
 import org.apache.coyote.http.common.HttpHeader;
@@ -18,15 +18,15 @@ import org.slf4j.LoggerFactory;
 import java.util.Map;
 import java.util.UUID;
 
-import static org.apache.coyote.handler.mapping.Path.MAIN;
-import static org.apache.coyote.handler.mapping.Path.UNAUTHORIZED;
+import static org.apache.coyote.handler.controller.Path.MAIN;
+import static org.apache.coyote.handler.controller.Path.UNAUTHORIZED;
 
-public class LoginMapping extends AbstractController {
+public class LoginController extends RequestController {
 
     private static final LoginManager loginManager = new SessionManager();
 
     private static final String TARGET_URI = "login";
-    private static final Logger log = LoggerFactory.getLogger(LoginMapping.class);
+    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
     @Override
     public boolean supports(final HttpRequest httpRequest) {
